requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.0
editdistance==0.6.2
epitran==1.24
et-xmlfile==1.1.0
idna==3.6
joblib==1.3.2
marisa-trie==1.1.0
munkres==1.1.4
numpy==1.26.3
openpyxl==3.1.2
pandas==2.2.0
panphon==0.20.0
pyarrow==15.0.0
pycryptodome==3.20.0
PyExecJS==1.5.1
pygame==2.5.2
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
tqdm==4.66.1
tzdata==2023.4
unicodecsv==0.14.1
urllib3==2.1.0
json_repair==0.25.3